Split out Yes/No, Radios, Checkboxes, Select and Autocomplete tests
